Drop renv
If the goal is to get the package on to CRAN, there's little point in using renv since CRAN requires compatibility with the latest versions of all dependencies. I also don't think this will cause a problem in practice, since will be relying on stable packages that are unlikely to have breaking API changes.