Fix handling of bare DATABRICKS_HOST URLs.
Some platforms set DATABRICKS_HOST without the leading "https" for
compatibility with existing ODBC code. Most other Databricks SDKs seem
to detect this automatically and prefix the host as required, but until
this commit this one did not.

(Note that this SDK is currently unusable inside Posit Workbench as a
result of this limitation.)

Closes #22.

Signed-off-by: Aaron Jacobs <aaron.jacobs@posit.co>